moved commands to separate file for organization
